requirements.txt
aiogram~=3.18.0
python-dotenv~=1.0.1
ollama~=0.4.7